requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-huggingface
langchain-text-splitters
chromadb
sentence-transformers
pypdf
openai
python-dotenv
